improved archiving

+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/dmd/DMDService.java
@@ -43,7 +43,7 @@ public class DMDService implements InitializingBean {
   public boolean delete(@NonNull Game game) {
     try {
       DMDPackage dmdPackage = getDMDPackage(game);
-      if (dmdPackage != null) {
+      if (dmdPackage != null && dmdPackage.isValid()) {
         File dir = getDmdFolder(game);
         if (dir.exists()) {
           FileUtils.deleteDirectory(dir);
@@ -59,9 +59,7 @@ public class DMDService implements InitializingBean {
 
   @Nullable
   public DMDPackage getDMDPackage(@NonNull Game game) {
-
     if (game.getDMDType() != null && !DMDPackageTypes.Standard.equals(game.getDMDType())) {
-
       DMDPackage dmdPackage = new DMDPackage();
       dmdPackage.setDmdPackageTypes(game.getDMDType());
